requirements/prod.txt
Django==5.0
python-dotenv==1.0.1
psycopg2==2.9.9
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1